Don't block on second run

+++ cleanup.go
@@ -2,6 +2,7 @@ package main
 
 import (
 	"context"
+	"errors"
 
 	"github.com/containerd/containerd"
 	"github.com/containerd/containerd/errdefs"
@@ -14,6 +15,13 @@ func cleanup(ctx context.Context, container containerd.Container) error {
 			return nil
 		}
 		return err
+	}
+	status, err := task.Status(ctx)
+	if err != nil {
+		return err
+	}
+	if status.Status != containerd.Stopped {
+		return errors.New("unable to start running container")
 	}
 	_, err = task.Delete(ctx)
 	return err
